+++ app/modules/debug/app/components/events.vue
@@ -17,15 +17,16 @@
                 <tbody>
                 <tr>
                     <th>Event name</th>
-                    <th>Listener</th>
                     <th>Priority</th>
+                    <th>Listener</th>
                 </tr>
                 <tr v-for="listener in called">
                     <td><code>{{ listener.event }}</code></td>
+                    <td>{{ listener.priority }}</td>
                     <td v-if="listener.type === 'Closure'">
                         <a :href="listener.link" v-if="listener.link">{{ listener.relative }}</a>
                         <span v-else>{{ listener.relative }}</span>
-                         {{ listener.line }} - {{ listener.endline }}
+                        ({{ listener.line }} - {{ listener.endline }})
                     </td>
                     <td v-if="listener.type === 'Function'">
                         <a :href="listener.link" v-if="listener.link">{{ listener.function }}</a>
@@ -35,7 +36,6 @@
                         <a :href="listener.link" v-if="listener.link"><abbr :title="listener.class">{{ listener.class | short }}</abbr>::{{ listener.method }}</a>
                         <span v-else><abbr :title="listener.class">{{ listener.class | short }}</abbr>::{{ listener.method }}</span>
                     </td>
-                    <td>{{ listener.priority }}</td>
                 </tr>
                 </tbody>
             </table>
@@ -48,15 +48,16 @@
                 <tbody>
                 <tr>
                     <th>Event name</th>
-                    <th>Listener</th>
                     <th>Priority</th>
+                    <th>Listener</th>
                 </tr>
                 <tr v-for="listener in notcalled">
                     <td><code>{{ listener.event }}</code></td>
+                    <td>{{ listener.priority }}</td>
                     <td v-if="listener.type === 'Closure'">
                         <a :href="listener.link" v-if="listener.link">{{ listener.relative }}</a>
                         <span v-else>{{ listener.relative }}</span>
-                        {{ listener.line }} - {{ listener.endline }}
+                        ({{ listener.line }} - {{ listener.endline }})
                     </td>
                     <td v-if="listener.type === 'Function'">
                         <a :href="listener.link" v-if="listener.link">{{ listener.function }}</a>
@@ -66,7 +67,6 @@
                         <a :href="listener.link" v-if="listener.link"><abbr :title="listener.class">{{ listener.class | short }}</abbr>::{{ listener.method }}</a>
                         <span v-else><abbr :title="listener.class">{{ listener.class | short }}</abbr>::{{ listener.method }}</span>
                     </td>
-                    <td>{{ listener.priority }}</td>
                 </tr>
                 </tbody>
             </table>